Create Team Strength Maximization
In this challenge, you'll be working with a unique team-building concept. A team starts with a single member identified by their initial ID. The next team member's ID is calculated by summing the factorials of the digits of the previous member's ID. The team continues to grow until it encounters a previously seen member's ID. Your goal is to find the maximum strength achievable by any team using this process.

The team strength is calculated by multiplying the number of members in the team by the highest ID in the team.

Input:

The function max_team_strength takes a single parameter n (1 <= n <= 10^6), which is the initial ID of the team member.

Output:

The function should return an integer representing the maximum possible team strength that can be achieved through the process described above.